Split out the VFDProtocol code into a separate file.
