+++ FluidNC/src/Spindles/VFD/SiemensV20Protocol.cpp
@@ -104,6 +104,8 @@ Take note that the serial interface use EVEN parity!
 
 #include "SiemensV20Protocol.h"
 
+#include "../VFDSpindle.h"
+
 #include <algorithm>  // std::max
 
 namespace Spindles {
@@ -167,7 +169,7 @@ namespace Spindles {
             data.msg[5] = ScaledFreq & 0xFF;
         }
 
-        VFDDetail::response_parser SiemensV20Protocol::initialization_sequence(int index, ModbusCommand& data) {
+        VFDProtocol::response_parser SiemensV20Protocol::initialization_sequence(int index, ModbusCommand& data) {
             // NOTE: data length is excluding the CRC16 checksum.
             data.tx_length = 6;
             data.rx_length = 5;
@@ -179,7 +181,7 @@ namespace Spindles {
             data.msg[4] = 0x00;
             data.msg[5] = 0x01;
 
-            return [](const uint8_t* response, VFDSpindle* vfd, VFDDetail* detail) -> bool {
+            return [](const uint8_t* response, VFDSpindle* vfd, VFDProtocol* detail) -> bool {
                 /*
             The VFD does not have any noticeable registers to set this information up programmatically
             For now - it is user set in the software but is a typical setup
@@ -201,7 +203,7 @@ namespace Spindles {
             };
         }
 
-        VFDDetail::response_parser SiemensV20Protocol::get_current_speed(ModbusCommand& data) {
+        VFDProtocol::response_parser SiemensV20Protocol::get_current_speed(ModbusCommand& data) {
             // NOTE: data length is excluding the CRC16 checksum.
             data.tx_length = 6;
             data.rx_length = 5;
@@ -213,7 +215,7 @@ namespace Spindles {
             data.msg[4] = 0x00;
             data.msg[5] = 0x01;
 
-            return [](const uint8_t* response, VFDSpindle* vfd, VFDDetail* detail) -> bool {
+            return [](const uint8_t* response, VFDSpindle* vfd, VFDProtocol* detail) -> bool {
                 auto    siemensV20      = static_cast<SiemensV20Protocol*>(detail);
                 int16_t Scaledfrequency = ((response[3] << 8) | response[4]);
                 int16_t frequency       = float(Scaledfrequency) / (-1 * (siemensV20->_FreqScaler));
