+++ FluidNC/src/Spindles/VFD/SiemensV20Protocol.h
@@ -4,11 +4,11 @@
 
 #pragma once
 
-#include "../VFDSpindle.h"
+#include "VFDProtocol.h"
 
 namespace Spindles {
     namespace VFD {
-        class SiemensV20Protocol : public VFDDetail {
+        class SiemensV20Protocol : public VFDProtocol {
         protected:
             uint16_t _minFrequency = 0;    // PD011: frequency lower limit. Normally 0.
             uint16_t _maxFrequency = 400;  // PD005: max frequency the VFD will allow. Normally 400.
